Minimize output from HG for mr --minimal status.
mr produces too much information on status for mercurial repositories.
It makes it quite difficult to process data quickly or automatically.

This patch modifies the implementation of the status commnand for
mercurial repositories if the flags --minimal or --quiet are enabled,
limiting the output to that of hg status.

Running mr --minimal status will ONLY report modifications to tracked
repositories now.